Conflicts:
	ideca2/ideca2-web/src/main/webapp/pages/admin/articleManagement.xhtml

+++ ideca2/ideca2-client/src/main/java/pdev/financialbrains/client/GUI/BackOfficeUser/PricingZeroCouponBond.java
@@ -130,7 +130,7 @@ public class PricingZeroCouponBond extends JFrame {
 			public void actionPerformed(ActionEvent arg0) {
 				FixedIncomeSecuritie f = new FixedIncomeSecuritie();
 				Float FaceValue=Float.parseFloat(tf_faceValue.getText());
-				Integer TimeToMaturity=Integer.parseInt(tf_time.getText());
+				Float TimeToMaturity=Float.parseFloat(tf_time.getText());
 				Float CurrentYield=Float.parseFloat(tf_rate.getText());
 				Float Price=FixedIncomeSecurityManagementDelegate.doPriceZeroCouponBond(FaceValue, TimeToMaturity, CurrentYield);
